iot2050: add helper function to convert gpio number to name
This function converts a GPIO number known to the OS into a name
provided by the GPIO subsystem.

This prepares for gpio-chardev support on the iot2050 platform and
should be replaced as soon as the pin definitions in
mraa_siemens_iot2050() have been adopted to use GPIO pin names
instead of GPIO pin numbers.

Signed-off-by: Benedikt Niedermayr <benedikt.niedermayr@siemens.com>